Refactor login_handler.py for non-headless Firefox execution
This commit removes the headless execution option from the webdriver.Firefox instantiation in login_handler.py. It enhances readability by renaming the instance variables "username" and "password" to "username_field" and "password_field", respectively. Furthermore, it gets rid of the auto-execution of the login process when the file runs, providing more flexibility and control.

Signed-off-by: Mathew Conder <cybermc.tech@gmail.com>